fix warnings

+++ src/media_tools/dash_client.c
@@ -803,7 +803,7 @@ static void gf_dash_group_timeline_setup(GF_MPD *mpd, GF_DASH_Group *group, u64 
 		}
 
 		if (group->nb_segments_in_rep && (group->download_segment_index + nb_segs_in_update > group->nb_segments_in_rep)) {
-			if (group->download_segment_index < group->nb_segments_in_rep) {
+			if (group->download_segment_index < (s32)group->nb_segments_in_rep) {
 
 			} else {
 				GF_LOG(GF_LOG_WARNING, GF_LOG_DASH, ("[DASH] Not enough segments (%d needed vs %d indicated) to reach period endTime indicated in MPD - ignoring MPD duration\n", nb_segs_in_update, group->nb_segments_in_rep - group->download_segment_index ));
